refactor: home 패키지에 포함된 클래스 이름 및 경로 수정
