+++ app/src/main/java/com/depromeet/sloth/ui/SlothFragmentFactory.kt
@@ -1,10 +1,10 @@
-package com.depromeet.sloth.ui.home
+package com.depromeet.sloth.ui
 
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.FragmentFactory
-import com.depromeet.sloth.ui.home.lessonlist.LessonListFragment
-import com.depromeet.sloth.ui.home.mypage.MypageFragment
-import com.depromeet.sloth.ui.home.today.TodayFragment
+import com.depromeet.sloth.ui.list.LessonListFragment
+import com.depromeet.sloth.ui.manage.ManageFragment
+import com.depromeet.sloth.ui.today.TodayFragment
 
 class SlothFragmentFactory : FragmentFactory() {
     override fun instantiate(classLoader: ClassLoader, className: String): Fragment {
@@ -15,8 +15,8 @@ class SlothFragmentFactory : FragmentFactory() {
             LessonListFragment::class.java.name -> {
                 return LessonListFragment()
             }
-            MypageFragment::class.java.name -> {
-                return MypageFragment()
+            ManageFragment::class.java.name -> {
+                return ManageFragment()
             }
         }
