+++ app/src/main/java/com/depromeet/sloth/ui/login/LoginActivity.kt
@@ -9,7 +9,7 @@ import android.widget.Button
 import com.depromeet.sloth.R
 import com.depromeet.sloth.data.PreferenceManager
 import com.depromeet.sloth.ui.base.BaseActivity
-import com.depromeet.sloth.ui.home.HomeActivity
+import com.depromeet.sloth.ui.HomeActivity
 
 class LoginActivity : BaseActivity<LoginViewModel, ActivityLoginBinding>() {
