+++ app/src/main/java/com/depromeet/sloth/ui/manage/WithdrawDialog.kt
@@ -1,4 +1,4 @@
-package com.depromeet.sloth.ui.home.mypage
+package com.depromeet.sloth.ui.manage
 
 import android.app.Dialog
 import android.content.Context
@@ -8,16 +8,16 @@ import android.view.Window
 import android.widget.Button
 import com.depromeet.sloth.R
 
-class LogoutDialog(context: Context) {
+class WithdrawDialog(context: Context) {
 
-    lateinit var listener: LogoutDialogClickedListener
-    lateinit var btnLogout: Button
+    lateinit var listener: WithdrawDialogClickedListener
+    lateinit var btnWithdraw: Button
     lateinit var btnCancel: Button
 
     private val dlg = Dialog(context)
 
-    interface LogoutDialogClickedListener {
-        fun onLogoutClicked()
+    interface WithdrawDialogClickedListener {
+        fun onWithdrawClicked()
     }
 
     fun start() {
@@ -25,15 +25,15 @@ class LogoutDialog(context: Context) {
         dlg.requestWindowFeature(Window.FEATURE_NO_TITLE)
         /*커스텀 다이얼로그 radius 적용*/
         dlg.window?.setBackgroundDrawable(ColorDrawable(Color.TRANSPARENT))
-        dlg.setContentView(R.layout.dialog_mypage_logout)
+        dlg.setContentView(R.layout.dialog_mypage_withdraw)
 
-        btnLogout = dlg.findViewById(R.id.btn_logout)
-        btnLogout.setOnClickListener {
-            listener.onLogoutClicked()
+        btnWithdraw = dlg.findViewById(R.id.btn_withdraw)
+        btnWithdraw.setOnClickListener {
+            listener.onWithdrawClicked()
             dlg.dismiss()
         }
 
-        btnCancel = dlg.findViewById(R.id.btn_logout_cancel)
+        btnCancel = dlg.findViewById(R.id.btn_withdraw_cancel)
         btnCancel.setOnClickListener {
             dlg.dismiss()
         }
