+++ app/src/main/java/com/depromeet/sloth/ui/today/HeaderAdapter.kt
@@ -1,4 +1,4 @@
-package com.depromeet.sloth.ui.home.today
+package com.depromeet.sloth.ui.today
 
 import android.view.LayoutInflater
 import android.view.View
