+++ app/src/main/java/com/depromeet/sloth/ui/today/TodayFragment.kt
@@ -1,4 +1,4 @@
-package com.depromeet.sloth.ui.home.today
+package com.depromeet.sloth.ui.today
 
 import android.content.Intent
 import android.os.Bundle
@@ -10,6 +10,10 @@ import com.depromeet.sloth.data.network.home.LessonState
 import com.depromeet.sloth.data.network.home.LessonTodayResponse
 import com.depromeet.sloth.data.network.home.LessonUpdateCountResponse
 import com.depromeet.sloth.databinding.FragmentTodayBinding
+import com.depromeet.sloth.ui.ForbiddenDialog
+import com.depromeet.sloth.ui.LessonItemDecoration
+import com.depromeet.sloth.ui.LessonViewModel
+import com.depromeet.sloth.ui.WaitDialog
 import com.depromeet.sloth.ui.base.BaseFragment
 import com.depromeet.sloth.ui.detail.LessonDetailActivity
 import com.depromeet.sloth.ui.home.*
