+++ app/src/main/res/layout/activity_home.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     android:background="#F5F5F5"
     android:fitsSystemWindows="true"
-    tools:context=".ui.home.HomeActivity">
+    tools:context=".ui.HomeActivity">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
