+++ app/src/main/res/layout/fragment_manage.xml
